Fix: Allow link actions to wrap to the next line when the browser window is too narrow.
They were always intended to wrap, but due to an oversight/lack of testing the styles that make the link URL non-wrapping also applied to row actions.

Note: It may seem that it would be simpler to just apply the no-wrap rule to .blc-link-url. However, that doesn't work in IE 11 (and possibly others). Apparently, wrapping styles must be set on the containing block.


git-svn-id: https://plugins.svn.wordpress.org/broken-link-checker/trunk@1033853 b8457f37-d9ea-0310-8a92-e5e31aec5664