Switch to YouTube API v3.
The old API version will be shut down on April 20. The new API is very different and doesn't provide as much information about video status as the old one did. For example, when a video is private, the v3 API will return the same response as when the video doesn't exist. As a result, error messages displayed by BLC will now be a little less informative. 


git-svn-id: https://plugins.svn.wordpress.org/broken-link-checker/trunk@1136097 b8457f37-d9ea-0310-8a92-e5e31aec5664